Split FeedbackPanel.tsx below the 500-line hard cap instead of resealing
Codex correctly called out the previous fix: resealing a brand-new file's
module-size ceiling at 654 lines defeats the ratchet's purpose during the
exact decomposition work it exists to drive, and the ratchet doc explicitly
lists "a new file over 500 lines" as forbidden regardless.

Split it for real instead of relocating the debt:
- composerAttachments.ts: the attachment state (uploads, sketches, the
  attach-menu open/close and outside-click handling) as a
  useComposerAttachments hook, with no JSX.
- CompactFeedbackComposer.tsx: the studio thread's inline reply-box render,
  which was already a distinct UI shape branched on `compact` rather than a
  variant of the same markup.
- FeedbackPanel.tsx keeps the shared state and the send/builder-switch
  logic, and is now 443 lines with no ceiling override needed.

All 83 SubmissionStatusView tests and the full 1,889-test web suite still
pass unchanged (no behavior change — this is a pure extraction), confirming
one condition the manual copy risked losing (`state !== 'sent'` gating the
route note in the compact composer) was carried over correctly.
