+++ apps/web/src/surfaces/studio/FeedbackPanel.tsx
@@ -10,16 +10,12 @@ import { SwitchToPlatformControl, SwitchToSelfControl } from './StudioConnectCar
 import { submitImprovement } from '../../studioApi.js';
 import { recordStudioStep } from '../../visitTelemetry.js';
 import { toBase64PngList } from '../../attachmentImages.js';
-import { SketchModal } from '../../SketchModal.js';
-import { useClampToViewport } from '../../useClampToViewport.js';
+import { useComposerAttachments } from './composerAttachments.js';
+import { CompactFeedbackComposer } from './CompactFeedbackComposer.js';
 import './status-feedback.css';
 import './status-composer.css';
 
 export type BuilderHandoffHandler = () => Promise<void | { pending?: boolean }> | void | { pending?: boolean };
-
-export type ComposerAttachment = { id: string; name: string; dataUrl: string };
-
-const MAX_COMPOSER_ATTACHMENTS = 4;
 
 const SENT_RECEIPT_MS = 4500;
 
@@ -106,32 +102,9 @@ export function FeedbackPanel({
   const [notice, setNotice] = useState<string | null>(null);
   const [builder, setBuilder] = useState<BuilderKind>(initialBuilder);
   const inputRef = useRef<HTMLTextAreaElement | null>(null);
-  const [attachments, setAttachments] = useState<ComposerAttachment[]>([]);
-  // FileReader work not yet landed in attachments — Send waits for it.
-  const [pendingAttachmentReads, setPendingAttachmentReads] = useState(0);
-  const [attachMenuOpen, setAttachMenuOpen] = useState(false);
-  const [isSketchOpen, setIsSketchOpen] = useState(false);
-  const fileInputRef = useRef<HTMLInputElement | null>(null);
-  const attachMenuRef = useRef<HTMLDivElement | null>(null);
-  const attachPanelRef = useClampToViewport<HTMLDivElement>(attachMenuOpen);
-
-  useEffect(() => {
-    if (!attachMenuOpen) return;
-    const onPointerDown = (event: MouseEvent) => {
-      if (attachMenuRef.current && !attachMenuRef.current.contains(event.target as Node)) {
-        setAttachMenuOpen(false);
-      }
-    };
-    const onKeyDown = (event: KeyboardEvent) => {
-      if (event.key === 'Escape') setAttachMenuOpen(false);
-    };
-    document.addEventListener('mousedown', onPointerDown);
-    document.addEventListener('keydown', onKeyDown);
-    return () => {
-      document.removeEventListener('mousedown', onPointerDown);
-      document.removeEventListener('keydown', onKeyDown);
-    };
-  }, [attachMenuOpen]);
+  const sending = state === 'sending';
+  const attachmentsApi = useComposerAttachments(sending);
+  const { attachments, pendingAttachmentReads, resetAttachments } = attachmentsApi;
 
   useEffect(() => {
     setBuilder(initialBuilder);
@@ -208,44 +181,6 @@ export function FeedbackPanel({
     input.style.height = `${Math.min(input.scrollHeight, 220)}px`;
   };
 
-  const handleAttachFiles = (files: FileList | File[]) => {
-    if (state === 'sending') return;
-    Array.from(files).forEach((file) => {
-      if (!file.type.startsWith('image/')) return;
-      setPendingAttachmentReads((count) => count + 1);
-      const reader = new FileReader();
-      const done = () => setPendingAttachmentReads((count) => count - 1);
-      reader.onload = (e) => {
-        const dataUrl = e.target?.result as string;
-        if (dataUrl) {
-          setAttachments((prev) =>
-            prev.length >= MAX_COMPOSER_ATTACHMENTS
-              ? prev
-              : [
-                  ...prev,
-                  { id: `file-${Date.now()}-${Math.random().toString(36).slice(2, 7)}`, name: file.name, dataUrl },
-                ],
-          );
-        }
-        done();
-      };
-      reader.onerror = done;
-      reader.readAsDataURL(file);
-    });
-  };
-
-  const handleSaveSketch = (dataUrl: string) => {
-    setAttachments((prev) =>
-      prev.length >= MAX_COMPOSER_ATTACHMENTS
-        ? prev
-        : [...prev, { id: `sketch-${Date.now()}`, name: `Sketch ${prev.length + 1}`, dataUrl }],
-    );
-  };
-
-  const removeAttachment = (id: string) => {
-    setAttachments((prev) => prev.filter((item) => item.id !== id));
-  };
-
   const send = async (requestedText: string = trimmed) => {
     const message = requestedText.trim();
     if (message.length < 10 || state === 'sending' || pendingAttachmentReads > 0) return;
@@ -301,7 +236,7 @@ export function FeedbackPanel({
       }
       setState('sent');
       setText('');
-      setAttachments([]);
+      resetAttachments();
       // Back to the CSS height rather than the height the sent message grew it to: an
       // empty box the size of the last paragraph is a leftover, not a state.
       if (inputRef.current) inputRef.current.style.height = '';
@@ -368,12 +303,12 @@ export function FeedbackPanel({
       value={effectiveBuilder}
       onChange={handleBuilderChange}
       canChange={chooseBuilder}
-      disabled={state === 'sending' || agentWorking}
+      disabled={sending || agentWorking}
       platformUnavailable={platformUnavailable}
     />
   ) : null;
   const activeSelfHandoff =
-    !chooseBuilder && effectiveBuilder === 'self' && onSwitchToPlatform && state !== 'sending' ? (
+    !chooseBuilder && effectiveBuilder === 'self' && onSwitchToPlatform && !sending ? (
       <SwitchToPlatformControl
         compact
         active
@@ -386,7 +321,7 @@ export function FeedbackPanel({
     !chooseBuilder && agentWorking && effectiveBuilder === 'platform' && Boolean(onSwitchToSelf) && !stopRequested;
   // Hide the switch-to-self badge while STOP covers the same action.
   const activePlatformHandoff =
-    !chooseBuilder && effectiveBuilder === 'platform' && onSwitchToSelf && !showStop && state !== 'sending' ? (
+    !chooseBuilder && effectiveBuilder === 'platform' && onSwitchToSelf && !showStop && !sending ? (
       <SwitchToSelfControl compact active onSwitchToSelf={onSwitchToSelf} pending={stopRequested} />
     ) : null;
   const stopAndSwitchToSelf = async () => {
@@ -434,180 +369,33 @@ export function FeedbackPanel({
   // Compact composer: field above, builder/send toolbar below.
   // Empty (`is-empty`): placeholder and send share one row.
   if (compact) {
-    const sending = state === 'sending';
-    const empty = text.length === 0;
     return (
-      <div
-        className={`status-feedback status-composer is-compact${empty ? ' is-empty' : ''}${sending ? ' is-sending' : ''}`}
-        aria-busy={sending || undefined}
-        onClick={(event) => {
-          // Clicking card chrome focuses the textarea; skip real controls.
-          const target = event.target;
-          if (!(target instanceof Element)) return;
-          if (target.closest('button, a, textarea, input, select, [role="button"]')) return;
-          inputRef.current?.focus();
+      <CompactFeedbackComposer
+        text={text}
+        onTextChange={(value) => {
+          setText(value);
+          if (state === 'sent') setState('idle');
         }}
-      >
-        {failureReason === 'gate_red' && !quickActionDismissed && !sending ? (
-          <div className="status-feedback-quick-actions">
-            <button type="button" className="status-feedback-quick-action" onClick={sendDebugCi}>
-              <PixelIcon name="wrench" size={12} />
-              {t('statusView.feedback.debugCi')}
-            </button>
-          </div>
-        ) : null}
-        {routeNoteKey && !sending && state !== 'sent' && !error && !notice ? (
-          <p className="status-feedback-route">{t(routeNoteKey)}</p>
-        ) : null}
-        <textarea
-          ref={inputRef}
-          className="status-feedback-input"
-          value={text}
-          onChange={(event) => {
-            setText(event.target.value);
-            autoGrow();
-            if (state === 'sent') setState('idle');
-          }}
-          onKeyDown={(event) => {
-            // Enter sends; Shift+Enter keeps a newline (same as RemixAsk).
-            // Skip while IME is composing — Enter confirms a candidate there.
-            const native = event.nativeEvent;
-            if (event.key !== 'Enter' || event.shiftKey || native.isComposing || native.keyCode === 229) {
-              return;
-            }
-            event.preventDefault();
-            void send();
-          }}
-          onPaste={(event) => {
-            const pastedImages = Array.from(event.clipboardData?.items ?? [])
-              .filter((item) => item.type.startsWith('image/'))
-              .map((item) => item.getAsFile())
-              .filter((file): file is File => file !== null);
-            if (pastedImages.length > 0) handleAttachFiles(pastedImages);
-          }}
-          placeholder={t(composerHintKey)}
-          aria-label={t(titleKey)}
-          rows={1}
-          maxLength={2000}
-          disabled={sending}
-        />
-        {attachments.length > 0 && (
-          <div className="status-composer-attachments">
-            {attachments.map((item) => (
-              <div key={item.id} className="status-composer-attachment-chip">
-                <img src={item.dataUrl} alt={item.name} className="status-composer-attachment-thumb" />
-                <button
-                  type="button"
-                  className="status-composer-attachment-remove"
-                  onClick={() => removeAttachment(item.id)}
-                  title={t('hero.removeAttachment')}
-                  disabled={sending}
-                >
-                  <PixelIcon name="close" size={11} />
-                </button>
-              </div>
-            ))}
-          </div>
-        )}
-        <div className="status-composer-toolbar">
-          <div className="status-composer-toolbar-left">
-            <div className="status-composer-attach" ref={attachMenuRef}>
-              <button
-                type="button"
-                className={`status-composer-attach-btn${attachMenuOpen ? ' is-open' : ''}`}
-                onClick={() => setAttachMenuOpen((open) => !open)}
-                title={t('hero.attachMenuAria')}
-                aria-label={t('hero.attachMenuAria')}
-                aria-expanded={attachMenuOpen}
-                aria-haspopup="menu"
-                disabled={sending || attachments.length >= MAX_COMPOSER_ATTACHMENTS}
-              >
-                <PixelIcon name="plus" size={15} />
-              </button>
-              {attachMenuOpen && !sending ? (
-                <div className="prompt-attach-menu" role="menu" aria-label={t('hero.attachMenu')} ref={attachPanelRef}>
-                  <button
-                    type="button"
-                    className="prompt-attach-item"
-                    role="menuitem"
-                    onClick={() => {
-                      setAttachMenuOpen(false);
-                      fileInputRef.current?.click();
-                    }}
-                  >
-                    <PixelIcon name="image" size={16} /> {t('hero.uploadImage')}
-                  </button>
-                  <button
-                    type="button"
-                    className="prompt-attach-item"
-                    role="menuitem"
-                    onClick={() => {
-                      setAttachMenuOpen(false);
-                      setIsSketchOpen(true);
-                    }}
-                  >
-                    <PixelIcon name="palette" size={16} /> {t('hero.drawSketch')}
-                  </button>
-                </div>
-              ) : null}
-              <input
-                ref={fileInputRef}
-                type="file"
-                accept="image/*"
-                multiple
-                className="hidden-file-input"
-                onChange={(event) => {
-                  if (event.target.files && event.target.files.length > 0) {
-                    handleAttachFiles(event.target.files);
-                    event.target.value = '';
-                  }
-                }}
-              />
-            </div>
-            {builderControls}
-          </div>
-          <div className="status-composer-toolbar-right">
-            {sending && !showStop ? (
-              <span className="status-feedback-sending" role="status">
-                {t('statusView.feedback.sending')}
-              </span>
-            ) : null}
-            {showStop ? (
-              <button
-                type="button"
-                className="status-composer-stop"
-                onClick={() => void stopAndSwitchToSelf()}
-                aria-label={t('statusView.feedback.stopTitle')}
-                title={t('statusView.feedback.stopTitle')}
-              >
-                <PixelIcon name="stop" size={13} />
-              </button>
-            ) : stopRequested ? null : (
-              <button
-                type="button"
-                className="primary-btn status-composer-send"
-                onClick={() => void send()}
-                disabled={sending || trimmed.length < 10 || pendingAttachmentReads > 0}
-                aria-label={sending ? t('statusView.feedback.sending') : t('statusView.feedback.submit')}
-                title={sending ? t('statusView.feedback.sending') : t('statusView.feedback.submit')}
-              >
-                {sending ? (
-                  <span className="status-composer-send-spinner" aria-hidden="true" />
-                ) : (
-                  <PixelIcon name="arrowRight" size={13} />
-                )}
-              </button>
-            )}
-          </div>
-        </div>
-        {/* Sending's indicator now sits inline next to Send, above */}
-        {error || notice ? (
-          <div className="status-feedback-actions">
-            {error ? <p className="error">{error}</p> : <p className="status-feedback-notice">{notice}</p>}
-          </div>
-        ) : null}
-        <SketchModal isOpen={isSketchOpen} onClose={() => setIsSketchOpen(false)} onSaveSketch={handleSaveSketch} />
-      </div>
+        inputRef={inputRef}
+        autoGrow={autoGrow}
+        onSend={() => void send()}
+        trimmed={trimmed}
+        sending={sending}
+        error={error}
+        notice={notice}
+        isSent={state === 'sent'}
+        failureReason={failureReason}
+        quickActionDismissed={quickActionDismissed}
+        onSendDebugCi={sendDebugCi}
+        routeNoteKey={routeNoteKey}
+        composerHintKey={composerHintKey}
+        titleKey={titleKey}
+        attachmentsApi={attachmentsApi}
+        builderControls={builderControls}
+        showStop={showStop}
+        stopRequested={stopRequested}
+        onStopAndSwitchToSelf={() => void stopAndSwitchToSelf()}
+      />
     );
   }
