+++ eslint-rules/comment-prose-baseline.json
@@ -562,6 +562,8 @@
     "apps/web/src/surfaces/studio/CodeSurface.tsx": 630,
     "apps/web/src/surfaces/studio/codeSurfaceApi.ts": 82,
     "apps/web/src/surfaces/studio/codeTokens.ts": 177,
+    "apps/web/src/surfaces/studio/CompactFeedbackComposer.tsx": 75,
+    "apps/web/src/surfaces/studio/composerAttachments.ts": 21,
     "apps/web/src/surfaces/studio/connectApi.ts": 231,
     "apps/web/src/surfaces/studio/CreatorStudioView.handoff.test.tsx": 291,
     "apps/web/src/surfaces/studio/CreatorStudioView.test.tsx": 858,
@@ -571,7 +573,7 @@
     "apps/web/src/surfaces/studio/editorJsonLiveDiff.ts": 92,
     "apps/web/src/surfaces/studio/EditorPanel.test.tsx": 16,
     "apps/web/src/surfaces/studio/EditorPanel.tsx": 644,
-    "apps/web/src/surfaces/studio/FeedbackPanel.tsx": 969,
+    "apps/web/src/surfaces/studio/FeedbackPanel.tsx": 939,
     "apps/web/src/surfaces/studio/studio-shell.css.test.ts": 42,
     "apps/web/src/surfaces/studio/StudioBuildHistory.test.tsx": 103,
     "apps/web/src/surfaces/studio/StudioBuildHistory.tsx": 27,
