+++ eslint-rules/module-size-baseline.json
@@ -832,7 +832,7 @@
     "apps/web/src/surfaces/studio/EditorSurface.test.tsx": 74,
     "apps/web/src/surfaces/studio/EditorSurface.tsx": 320,
     "apps/web/src/surfaces/studio/editorSurfaceMode.ts": 16,
-    "apps/web/src/surfaces/studio/FeedbackPanel.tsx": 655,
+    "apps/web/src/surfaces/studio/FeedbackPanel.tsx": 443,
     "apps/web/src/surfaces/studio/NumberScrubber.test.tsx": 98,
     "apps/web/src/surfaces/studio/NumberScrubber.tsx": 87,
     "apps/web/src/surfaces/studio/roundSealed.test.ts": 22,
